backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
alembic==1.12.1
httpx==0.25.2
numpy==1.26.2
pandas==2.1.3
scikit-learn==1.3.2
joblib==1.3.2
scipy==1.11.4
scikit-optimize==0.9.0
statsmodels==0.14.0
lightgbm==4.1.0
xgboost==2.0.3
tensorflow==2.15.0
keras==2.15.0
pulp==2.7.0
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
selenium==4.15.2
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.23
apscheduler==3.10.4
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
rapidfuzz==3.5.2
psutil==5.9.6
pyarrow==14.0.1
matplotlib==3.8.2
ruff==0.14.13
mypy==1.19.0
pytest==7.4.3
pytest-asyncio==0.21.1
hypothesis==6.92.1